Update https_with_nginx.md
Added values to have the trash page not timeout on the browser site. Browser doesn't wait that long otherwise and will show a white page or a timeout window. Had to fiddle a while to get it working with these values.
"Buffering off" should be set for seafdav as well as it may timeout the same way as the webgui does.