add python-requests dependency
If a user wishes to install following these instructions and does not setup nginx/apache, then they will not be able to access the web interface.  If you follow the current instructions you will get an HTTP 500 error when loading the seafile webui at HOSTNAME:8000, this is because there is no webserver installed, not even a minimal one like python-requests.  I propose that following the barebones instructions should install at least a minimal webui.